clarified diff of two servers
Added clarification that diff can be done between two servers (from "from" and "to"). 

The current listed examples presumed only to compare a current commandbox server "to" another.

And while the others use "anotherservername", I changed the new line to use "oneserver" and "another", like is used on the "transfer" example right above it (and since the from and to values could be paths to the server settings home, rather than just the name of another commandbox server).